fix: preservar semântica de avaliados no IDEB 2025
